requirements.txt
google-generativeai>=0.3.0
langgraph>=0.0.10
langchain-core>=0.0.10
requests>=2.25.0
pylint>=2.8.0
pytest>=6.0.0
pytest-cov>=2.12.0
python-dotenv>=0.15.0